move new template tests from test_floris.py to test_templates.py
